Refactor vl.Source(ds)
