+++ cartoframes/__init__.py
@@ -2,5 +2,5 @@ from .context import CartoContext
 from .credentials import Credentials
 from .layer import BaseMap, QueryLayer, Layer
 from .styling import BinMethod
-from .datasets import Dataset
+from .dataset import Dataset
 from .__version__ import __version__
