--- cartoframes/carto_vl/sources.py
@@ -1,9 +1,0 @@
-from .source.dataset import Dataset
-from .source.sql import SQL
-from .source.geojson import GeoJSON
-
-
-class Sources():
-    Dataset = Dataset
-    SQL = SQL
-    GeoJSON = GeoJSON
