+++ cartoframes/context.py
@@ -31,7 +31,7 @@ from .maps import (non_basemap_layers, get_map_name,
                    get_map_template, top_basemap_layer_url)
 from .analysis import Table
 from .__version__ import __version__
-from .datasets import Dataset, recursive_read, postprocess_dataframe, get_columns
+from .dataset import Dataset, recursive_read, postprocess_dataframe, get_columns
 
 if sys.version_info >= (3, 0):
     from urllib.parse import urlparse, urlencode
